Fixed logging issue
When a task was logging at a level below the root logger it would not
record the log messages because the root logger was rejecting them
before our task's handler could get them.  In such a scenario, we now
temporarily lower the root logger's log level to the one we want to log
at and then revert it.

WARNING: This only applies to the root level.  If a library specifically
set its log level then the same problem will occur as it will be
intercepted by its logger before it even reaches the root.  In which
case the standard way of changing that libraries log level should be
used.