Set explicit PyYAML version and fixe test
PyYAML 5.4 introduced some change that means that FullLoader can't load
bson.ObjectIDs (while Loader can).  Unfortunately it's not easy to fix
this is a general way so for now I'm pinning the max PyYAML at 5.3.1.

Fixed a mistake in test which should have applied to `minki project
list` but was actually part of `minki list` test.